삭제, 수정시 확인 취소창, 추가버튼을 등록버튼으로 변경

+++ client/src/views/proc/procInfo.vue
@@ -4,7 +4,7 @@
     <!-- 우측 버튼 모음 영역 -->
     <div class="d-flex justify-content-between mb-3">
       <div> <!-- 버튼 왼쪽 정렬 -->
-        <button class="btn btn-primary me-2" @click="addProc">추가</button>
+        <button class="btn btn-primary me-2" @click="addProc">등록</button>
         <button class="btn btn-danger" @click="deleteProc(proc_no)">삭제</button>
       </div>
       <div> <!-- 버튼 오른쪽 정렬 -->
@@ -203,7 +203,9 @@ export default {
         this.getProcList(this.proc_no);
         return;
       } else {
-        this.procUpdate(); // 수정내용 저장
+        if (confirm('정말로 수정하시겠습니까?\n변경된 내용은 즉시 적용됩니다.')) {
+          this.procUpdate(); // 수정내용 저장
+        }
       }
     },
     // 추가 버튼 클릭시 실행할 함수
@@ -223,15 +225,17 @@ export default {
         return;
       } else {
         if (procNo != null) { // 선택된 proc가 있을 경우 
-          let result = await axios.delete(`/api/proc/${procNo}`)
-            .catch(err => console.log(err));
-          let sqlRes = result.data;
-          let sqlResult = sqlRes.affectedRows;
-          if (sqlResult > 0) {
-            alert('정상적으로 삭제되었습니다.');
-            this.$emit('proc-reload');
-          } else {
-            alert('삭제되지 않았습니다.');
+          if (confirm('정말로 삭제하시겠습니까?\n삭제된 데이터는 복구할 수 없습니다.')) {
+            let result = await axios.delete(`/api/proc/${procNo}`)
+              .catch(err => console.log(err));
+            let sqlRes = result.data;
+            let sqlResult = sqlRes.affectedRows;
+            if (sqlResult > 0) {
+              alert('정상적으로 삭제되었습니다.');
+              this.$emit('proc-reload');
+            } else {
+              alert('삭제되지 않았습니다.');
+            }
           }
         } else { // 선택된 proc이 없을 경우
           alert("삭제할 제품공정을 선택하세요");
